Added models and tweaked schemas where needed. - Changed inconsistent variable naming in schemas. - Added all models and eloquent relations to new activity features.

+++ database/migrations/2021_10_25_000001_create_activity_details_table.php
@@ -19,7 +19,7 @@ class CreateActivityDetailsTable extends Migration
             $table->foreignId('resource_id')->constrained()->nullable();
             $table->string('name');
             $table->text('instructions');
-            $table->integer('minutes_to_complete');
+            $table->unsignedSmallInteger('minutes_to_complete');
             $table->timestamps();
         });
     }
